+++ src/models/User.ts
@@ -1,10 +1,18 @@
-
+export enum USER_ROLES {
+    NORMAL = "NORMAL",
+    ADMIN = "ADMIN"
+} 
+export interface TokenPayload {
+    id: string,
+    name: string,
+    role: USER_ROLES
+} 
 export interface UserDB{
     id: string,
     name: string,
     email: string,
     password: string,
-    role: string,
+    role: USER_ROLES,
     created_at: string
 }
 
@@ -12,7 +20,7 @@ export interface UserModel{
     id: string,
     name: string,
     email: string,
-    role: string,
+    role: USER_ROLES,
     createdAt: string
 }
 
@@ -22,40 +30,49 @@ export class User{
         private name: string,
         private email: string,
         private password: string,
-        private role: string,
+        private role: USER_ROLES,
         private createdAt: string
     ){}
+
     public getId(): string{
         return this.id
-    }    
+    }
+    public setId(value: string): void{
+        this.id = value
+    } 
+    
     public getName(): string{
         return this.name
     }
     public setName(value: string): void{
         this.name = value
     }
+
     public getEmail(): string{
         return this.email
     }
     public setEmail(value: string): void{
         this.email = value
     }
+
     public getPassword(): string{
         return this.password
     }
     public setPassword(value: string): void{
         this.password = value
     }
-    public getRole(): string{
+
+    public getRole(): USER_ROLES{
         return this.role
     }
-    public setRole(value: string): void{
+    public setRole(value: USER_ROLES): void{
         this.role = value
     }
+
     public getCreatedAt(): string{
         return this.createdAt
     } 
-    
+    // vai retornar um userDb
     public toDBModel(): UserDB{
         return{
             id: this.id,
