+++ src/services/TokenManager.ts
@@ -1,18 +1,11 @@
 import jwt from 'jsonwebtoken'
 import dotenv from 'dotenv'
+import { TokenPayload } from '../models/User'
 
 dotenv.config()
 
-export enum USER_ROLES {
-    NORMAL = "NORMAL",
-    ADMIN = "ADMIN"
-} // realocar para o arquivo da entidade User
 
-export interface TokenPayload {
-    id: string,
-    name: string,
-    role: USER_ROLES
-} // realocar para o arquivo da entidade User
+
 
 export class TokenManager {
     public createToken = (payload: TokenPayload): string => {
